Make getCommonsChunkFilename a util
